backend/requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.8
django-cors-headers==4.9.0
django-filter==25.2
django-seed==0.3.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
Faker==38.2.0
idna==3.11
inflection==0.5.1
packaging==25.0
pillow==12.0.0
PyJWT==2.10.1
pyotp==2.9.0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
sqlparse==0.5.3
toposort==1.10
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
